도장뷰에도 스크롤뷰 추가 + 내 일기에 도장반영

+++ app/src/main/java/com/example/drwaing/data/diary/Stamp.kt
@@ -9,7 +9,7 @@ data class Stamp(
     @SerializedName("stampType")
     val stampType: String,
     @SerializedName("x")
-    val x: Int,
+    val x: Double,
     @SerializedName("y")
-    val y: Int
+    val y: Double
 )
